crud job succesful

+++ src/main/java/com/carpentery/springboot/app/repository/CategoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.carpentery.springboot.app.model.CategoryModel;
 
 
-public interface CategoryRepository extends JpaRepository<CategoryModel, Integer> {
+public interface CategoryRepository extends JpaRepository<CategoryModel, Long> {
 }
