+++ src/main/java/com/carpentery/springboot/app/repository/TipoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.carpentery.springboot.app.model.TipoModel;
 
-public interface TipoRepository extends JpaRepository<TipoModel, Integer> {
+public interface TipoRepository extends JpaRepository<TipoModel, Long> {
 }
